requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
streamlit
psycopg2-binary
pandas
plotly
vanna
openai